Add method docs and some small refactors
