+++ services/openstad/importService.js
@@ -2,8 +2,15 @@ const tar = require('tar');
 const fs = require('fs').promises;
 const fetch = require('node-fetch');
 
-// check if fileUrl isset then download the file
-// otherwise assume file is upload and use multer to process local upload
+/**
+ * Handle and extract import file
+ * - check if fileUrl isset then download the file
+ * - otherwise assume file is upload and use multer to process local upload
+ * @param newSite
+ * @param file
+ * @param fileUrl
+ * @returns {Promise<*>}
+ */
 exports.handleImportFile = async (newSite, file, fileUrl) => {
 
   if (fileUrl) {
