+++ services/openstad/models/siteData.js
@@ -1,0 +1,5 @@
+module.exports = function SiteData(site, choiceGuides, cmsAttachments, mongoPath, oauthClients) {
+  this.apiData = {site, choiceGuides};
+  this.cmsData = {attachments: cmsAttachments, mongoPath};
+  this.oauthData = {clients: oauthClients};
+};
